requirements.txt
Alembic>=1.12.0
Flask-Migrate>=4.0.5
Flask-SQLAlchemy>=3.1.0
Flask-WTF==1.2.1
Flask==3.0.0
Pillow==10.1.0
SQLALchemy>=2.0.21
gunicorn==21.2.0
honcho==1.1.0
playwright>=1.31.1